refactor: 패키지 정리 및 README에 chapter2 추가
