--- src/main/java/com/perenok/objects/chapter2/Customer.java
@@ -1,4 +1,0 @@
-package com.perenok.objects.chapter2;
-
-public class Customer {
-}
