--- src/main/java/com/perenok/objects/chapter2/DiscountCondition.java
@@ -1,5 +1,0 @@
-package com.perenok.objects.chapter2;
-
-public interface DiscountCondition {
-    boolean isSatisfiedBy(Screening screening);
-}
