--- src/main/java/com/perenok/objects/chapter2/NoneDiscountPolicy.java
@@ -1,9 +1,0 @@
-package com.perenok.objects.chapter2;
-
-public class NoneDiscountPolicy extends DiscountPolicy {
-
-    @Override
-    protected Money getDiscountAmount(Screening screening) {
-        return Money.ZERO;
-    }
-}
