+++ src/main/java/com/perenok/objects/chapter2/movie/Customer.java
@@ -1,0 +1,4 @@
+package com.perenok.objects.chapter2.movie;
+
+public class Customer {
+}
