+++ src/main/java/com/perenok/objects/chapter2/movie/Movie.java
@@ -1,4 +1,6 @@
-package com.perenok.objects.chapter2;
+package com.perenok.objects.chapter2.movie;
+
+import com.perenok.objects.chapter2.policy.DefaultDiscountPolicy;
 
 import java.time.Duration;
 
@@ -7,9 +9,9 @@ public class Movie {
     private String title;
     private Duration runningTime;
     private Money fee;
-    private DiscountPolicy discountPolicy;
+    private DefaultDiscountPolicy discountPolicy;
 
-    public Movie(String title, Duration runningTime, Money fee, DiscountPolicy discountPolicy) {
+    public Movie(String title, Duration runningTime, Money fee, DefaultDiscountPolicy discountPolicy) {
         this.title = title;
         this.runningTime = runningTime;
         this.fee = fee;
@@ -17,7 +19,7 @@ public class Movie {
     }
 
     public Money getFee() {
-        return null;
+        return fee;
     }
 
     public Money calculateMovieFee(Screening screening) {
