+++ src/main/java/com/perenok/objects/chapter2/movie/Screening.java
@@ -1,4 +1,4 @@
-package com.perenok.objects.chapter2;
+package com.perenok.objects.chapter2.movie;
 
 import java.time.LocalDateTime;
