+++ src/main/java/com/perenok/objects/chapter2/policy/DefaultDiscountPolicy.java
@@ -1,16 +1,20 @@
-package com.perenok.objects.chapter2;
+package com.perenok.objects.chapter2.policy;
+
+import com.perenok.objects.chapter2.movie.Money;
+import com.perenok.objects.chapter2.movie.Screening;
 
 import java.util.Arrays;
 import java.util.List;
 
-public abstract class DiscountPolicy {
+public abstract class DefaultDiscountPolicy implements DiscountPolicy {
 
     private List<DiscountCondition> conditions;
 
-    public DiscountPolicy(DiscountCondition... conditions) {
+    public DefaultDiscountPolicy(DiscountCondition... conditions) {
         this.conditions = Arrays.asList(conditions);
     }
 
+    @Override
     public Money calculateDiscountAmount(Screening screening) {
         for (DiscountCondition condition : conditions) {
             if (condition.isSatisfiedBy(screening)) {
