+++ src/main/java/com/perenok/objects/chapter2/policy/DiscountCondition.java
@@ -1,0 +1,7 @@
+package com.perenok.objects.chapter2.policy;
+
+import com.perenok.objects.chapter2.movie.Screening;
+
+public interface DiscountCondition {
+    boolean isSatisfiedBy(Screening screening);
+}
