+++ src/main/java/com/perenok/objects/chapter2/policy/SequenceCondition.java
@@ -1,4 +1,6 @@
-package com.perenok.objects.chapter2;
+package com.perenok.objects.chapter2.policy;
+
+import com.perenok.objects.chapter2.movie.Screening;
 
 public class SequenceCondition implements DiscountCondition {
